move useClose composable
