--- src/components/GroupManager/composables/useClose.ts
@@ -1,12 +1,0 @@
-import { constructChannelPath } from '/@/router'
-import { useRouter } from 'vue-router'
-import { useBrowserSettings } from '/@/store/app/browserSettings'
-
-const useClose = () => {
-  const router = useRouter()
-  const { lastOpenChannelName } = useBrowserSettings()
-  const close = () =>
-    router.push(constructChannelPath(lastOpenChannelName.value))
-  return { close }
-}
-export default useClose
