+++ src/components/Settings/MobileTabFrame.vue
@@ -15,9 +15,11 @@
 import TabContentTitle from './TabContentTitle.vue'
 import ReturnButton from '/@/components/UI/ReturnButton.vue'
 import CloseButton from '/@/components/UI/CloseButton.vue'
+import useClose from '/@/composables/useClose'
 import useSettingsNavigation from './composables/useNavigation'
 
-const { close, showRoot } = useSettingsNavigation()
+const { showRoot } = useSettingsNavigation()
+const { close } = useClose()
 </script>
 
 <style lang="scss" module>
