+++ src/components/Settings/MobileTabSelector.vue
@@ -23,13 +23,14 @@ import CloseButton from '/@/components/UI/CloseButton.vue'
 import VersionInfo from '/@/components/UI/VersionInfo.vue'
 import MobileTabSelectorItem from './MobileTabSelectorItem.vue'
 import SafariWarning from './SafariWarning.vue'
+import useClose from '/@/composables/useClose'
 import { computed } from 'vue'
 import { useRoute } from 'vue-router'
-import useSettingsNavigation, { navigations } from './composables/useNavigation'
+import { navigations } from './composables/useNavigation'
 
 const route = useRoute()
 const currentRouteName = computed(() => route.name)
-const { close } = useSettingsNavigation()
+const { close } = useClose()
 </script>
 
 <style lang="scss" module>
